th8: add phase warps for Mokou last spell
th7: fix warning related to everlasting bgm, in spite of the previous
commit message's best wishes

+++ thprac/src/thprac/thprac_th07.cpp
@@ -1862,7 +1862,7 @@ namespace TH07 {
         }
 
         if (opcode == AUDIO_START) {
-            if (arg != o.cur_song) {
+            if (arg != (uint32_t)o.cur_song) {
                 o.cur_song = arg;
             }
             else {
